main/requirements.txt
discord.py==2.2.2
python-dotenv==0.21.1
revChatGPT==1.2.1
pytube==12.1.3
moviepy==1.0.3
PyNaCl==1.5.0
pygame==2.3.0
youtube-dl==2021.12.17